[Fix]More split command files
