+++ command/accessLocalFile.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"isso0424/spotify_CLI/command/file"
 	"isso0424/spotify_CLI/command/parse"
+	"isso0424/spotify_CLI/command/request"
 	"isso0424/spotify_CLI/selfmadetypes"
 	"isso0424/spotify_CLI/util"
 )
@@ -29,23 +30,13 @@ func (cmd save) Execute() (err error) {
 		return
 	}
 
-	if checkDuplicateName(name, playlistList) {
+	if util.CheckDuplicateName(name, playlistList) {
 		err = file.SavePlayList(list)
 	} else {
 		err = &selfmadetypes.NameDuplicateError{Target: name}
 	}
 
 	return
-}
-
-func checkDuplicateName(name string, playlistList []selfmadetypes.PlayList) bool {
-	for _, content := range playlistList {
-		if content.Name == name {
-			return false
-		}
-	}
-
-	return true
 }
 
 // Execute is excution command function.
@@ -64,6 +55,43 @@ func (cmd show) Execute() (err error) {
 			target.URI,
 		)
 	}
+
+	return
+}
+
+// Execute is excution command function.
+func (cmd random) Execute(token *string) (err error) {
+	playlists, err := file.LoadPlayList()
+	if err != nil {
+		return
+	}
+
+	choisePlaylist := util.Choose(playlists)
+	err = request.PlayFromURL(token, choisePlaylist.URI)
+
+	return
+}
+
+// Execute is excution command function.
+func (cmd load) Execute(token *string) (err error) {
+	var name string
+	util.Input("please input playlist name", "PlayListName", &name)
+
+	playlistList, err := file.LoadPlayList()
+
+	if err != nil {
+		return
+	}
+
+	for _, target := range playlistList {
+		if target.Name == name {
+			fmt.Printf("play %s\n", target.Name)
+			err = request.PlayFromURL(token, target.URI)
+			return
+		}
+	}
+
+	err = &selfmadetypes.NotFound{Target: name}
 
 	return
 }
