+++ command/command_test.go
@@ -2,6 +2,7 @@ package command
 
 import (
 	"isso0424/spotify_CLI/selfmadetypes"
+	"isso0424/spotify_CLI/util"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -10,17 +11,17 @@ import (
 // TestSwitchRepeat is test function for switchRepeat()
 func TestSwitchRepeat(t *testing.T) {
 	var result string
-	result = switchRepeatState("off")
+	result = util.SwitchRepeatState("off")
 	assert.Equal(t, result, "context")
 
-	result = switchRepeatState("track")
+	result = util.SwitchRepeatState("track")
 	assert.Equal(t, result, "off")
 
-	result = switchRepeatState("context")
+	result = util.SwitchRepeatState("context")
 	assert.Equal(t, result, "track")
 }
 
-// TestCheckDuplicateName is test function for checkDuplicateName()
+// TestCheckDuplicateName is test function for CheckDuplicateName()
 func TestCheckDuplicateName(t *testing.T) {
 	var playlistList []selfmadetypes.PlayList
 	var result bool
@@ -35,9 +36,9 @@ func TestCheckDuplicateName(t *testing.T) {
 		},
 	}
 
-	result = checkDuplicateName("playlist", playlistList)
+	result = util.CheckDuplicateName("playlist", playlistList)
 	assert.Equal(t, result, false)
 
-	result = checkDuplicateName("not found", playlistList)
+	result = util.CheckDuplicateName("not found", playlistList)
 	assert.Equal(t, result, true)
 }
