+++ command/other.go
@@ -1,0 +1,39 @@
+package command
+
+import (
+	"encoding/json"
+	"fmt"
+	"isso0424/spotify_CLI/auth"
+	"isso0424/spotify_CLI/command/request"
+	"isso0424/spotify_CLI/selfmadetypes"
+)
+
+// Execute is excution command function.
+func (cmd welcome) Execute(token *string) (err error) {
+	response, _, err := request.CreateRequest(token, selfmadetypes.GET, "/me", nil)
+	if err != nil {
+		return
+	}
+
+	var userInfo selfmadetypes.User
+	err = json.Unmarshal(response, &userInfo)
+	if err != nil {
+		return
+	}
+
+	fmt.Printf("ようこそ! %sさん!\n", userInfo.DisplayName)
+
+	return
+}
+
+// Execute is excution command function.
+func (cmd refresh) Execute(token *string) error {
+	tokenPtr, err := auth.GetToken()
+	if err != nil {
+		return err
+	}
+
+	*token = *tokenPtr
+
+	return nil
+}
